+++ tests/res/stubber/expected/with_complex_lists.py
@@ -4,7 +4,6 @@ __all__ = [
     'CfgBarYomamaStub',
     'CfgBarStub',
     'CfgStub',
-    'AlvissConfigStub',
 ]
 
 from typing import *
@@ -13,30 +12,26 @@ from alviss.structs.cfgstub import _BaseCfgStub
 from alviss.structs import BaseConfig
 
 
-class CfgFooStub(_BaseCfgStub):
+class CfgFooStub(_BaseCfgStub, dict):
     my_list: Union[List[str], Empty]
     my_required_list: List[str]
 
 
-class CfgFoolStub(_BaseCfgStub):
+class CfgFoolStub(_BaseCfgStub, dict):
     myDifferentList: Union[List[str], Empty]
     myDifferentMultiTypeList: Union[List[Union[str, float]], Empty]
 
 
-class CfgBarYomamaStub(_BaseCfgStub):
+class CfgBarYomamaStub(_BaseCfgStub, dict):
     myDifferentRequiredList: List[Union[str, int]]
     myDifferentRequiredListByKeyName: List[Any]
 
 
-class CfgBarStub(_BaseCfgStub):
+class CfgBarStub(_BaseCfgStub, dict):
     yomama: CfgBarYomamaStub
 
 
-class CfgStub(_BaseCfgStub):
+class CfgStub(_BaseCfgStub, dict):
     Foo: CfgFooStub
     fool: Union[CfgFoolStub, Empty]
     bar: CfgBarStub
-
-
-class AlvissConfigStub(BaseConfig, CfgStub):
-    pass