+++ tests/res/stubber/expected/with_complex_maps_and_lists.py
@@ -6,7 +6,7 @@ __all__ = [
     'CfgHereBeStufMandatoryPeopleStub',
     'CfgHereBeStufStub',
     'CfgStub',
-    'AlvissConfigStub',
+    'ComplexConfig',
 ]
 
 from typing import *
@@ -15,42 +15,42 @@ from alviss.structs.cfgstub import _BaseCfgStub
 from alviss.structs import BaseConfig
 
 
-class CfgImportTestSkillsStub(_BaseCfgStub):
+class CfgImportTestSkillsStub(_BaseCfgStub, dict):
     level: Union[int, Empty]
     group: str
 
 
-class CfgImportTestInventoryStub(_BaseCfgStub):
+class CfgImportTestInventoryStub(_BaseCfgStub, dict):
     weight: Union[float, Empty]
     quantity: Union[int, Empty]
 
 
-class CfgImportTestStub(_BaseCfgStub):
+class CfgImportTestStub(_BaseCfgStub, dict):
     other_list: Union[List[Union[str, float]], Empty]
     required_list: List[str]
     skills: Dict[str, CfgImportTestSkillsStub]
     inventory: Union[Dict[str, CfgImportTestInventoryStub], Empty]
 
 
-class CfgHereBeStufPeopleStub(_BaseCfgStub):
+class CfgHereBeStufPeopleStub(_BaseCfgStub, dict):
     name: str
     age: Union[int, Empty]
     favorite_color: Union[str, Empty]
 
 
-class CfgHereBeStufMandatoryPeopleStub(_BaseCfgStub):
+class CfgHereBeStufMandatoryPeopleStub(_BaseCfgStub, dict):
     name: str
     age: Union[int, Empty]
     level: int
 
 
-class CfgHereBeStufStub(_BaseCfgStub):
+class CfgHereBeStufStub(_BaseCfgStub, dict):
     people: List[CfgHereBeStufPeopleStub]
     mandatory_people: List[CfgHereBeStufMandatoryPeopleStub]
     sub_seven: Dict[str, Any]
 
 
-class CfgStub(_BaseCfgStub):
+class CfgStub(_BaseCfgStub, dict):
     my_list: Union[List[str], Empty]
     my_required_list: List[str]
     myDifferentList: Union[List[str], Empty]
@@ -60,5 +60,5 @@ class CfgStub(_BaseCfgStub):
     here_be_stuf: CfgHereBeStufStub
 
 
-class AlvissConfigStub(BaseConfig, CfgStub):
+class ComplexConfig(BaseConfig, CfgStub):
     pass